feat: durbar71 SMS, auto-approve butchers, decimal prices, full-bleed landing
- SMS: switch provider to durbar71.com (POST userId/password); V17 migration + SmsService/AdminController

- Butchers: auto-approve on registration (status APPROVED instead of PENDING)

- Pricing: cow_price/goat_price now NUMERIC(10,2) / BigDecimal end-to-end for percentage pricing (e.g. 13.5); V18 migration; forms use parseFloat + step

- Landing: full-bleed object-contain hero (no side strips/top gap, buttons always visible, no scroll)

- Add previously-untracked SMS migrations V15, V16

- Merge frontend into the main repo (remove nested gitlink) so it is one monorepo

+++ backend/src/main/java/com/kosaibari/domain/Butcher.java
@@ -12,9 +12,9 @@ public class Butcher {
     private UUID userId;
     private String whatsapp;
     private String photoUrl;
-    private Integer cowPrice;
+    private BigDecimal cowPrice;
     private PriceType cowPriceType;
-    private Integer goatPrice;
+    private BigDecimal goatPrice;
     private PriceType goatPriceType;
     private Integer cowCapacity;
     private Integer goatCapacity;
